Created API and files to deploy to Heroku to test API
