+++ requirements.txt
@@ -1,0 +1,87 @@
+altgraph==0.10.2
+aniso8601==8.1.0
+asn1crypto==0.24.0
+bdist-mpkg==0.5.0
+bonjour-py==0.3
+cachetools==3.1.1
+certifi==2020.12.5
+cffi==1.14.3
+chardet==4.0.0
+click==7.1.2
+cryptography==2.6.1
+enum34==1.1.6
+Flask==1.1.2
+Flask-Cors==3.0.9
+Flask-RESTful==0.3.8
+future==0.17.1
+futures==3.3.0
+google-api-core==1.24.1
+google-api-python-client==1.12.8
+google-auth==1.24.0
+google-auth-httplib2==0.0.4
+google-auth-oauthlib==0.4.1
+googleapis-common-protos==1.52.0
+httplib2==0.18.1
+idna==2.10
+ipaddress==1.0.22
+itsdangerous==1.1.0
+Jinja2==2.11.2
+macholib==1.5.1
+MarkupSafe==1.1.1
+matplotlib==1.3.1
+modulegraph==0.10.4
+numpy==1.8.0rc1
+oauthlib==3.1.0
+protobuf==3.14.0
+py2app==0.7.3
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.19
+pyobjc-core==2.5.1
+pyobjc-framework-Accounts==2.5.1
+pyobjc-framework-AddressBook==2.5.1
+pyobjc-framework-AppleScriptKit==2.5.1
+pyobjc-framework-AppleScriptObjC==2.5.1
+pyobjc-framework-Automator==2.5.1
+pyobjc-framework-CFNetwork==2.5.1
+pyobjc-framework-Cocoa==2.5.1
+pyobjc-framework-Collaboration==2.5.1
+pyobjc-framework-CoreData==2.5.1
+pyobjc-framework-CoreLocation==2.5.1
+pyobjc-framework-CoreText==2.5.1
+pyobjc-framework-DictionaryServices==2.5.1
+pyobjc-framework-EventKit==2.5.1
+pyobjc-framework-ExceptionHandling==2.5.1
+pyobjc-framework-FSEvents==2.5.1
+pyobjc-framework-InputMethodKit==2.5.1
+pyobjc-framework-InstallerPlugins==2.5.1
+pyobjc-framework-InstantMessage==2.5.1
+pyobjc-framework-LatentSemanticMapping==2.5.1
+pyobjc-framework-LaunchServices==2.5.1
+pyobjc-framework-Message==2.5.1
+pyobjc-framework-OpenDirectory==2.5.1
+pyobjc-framework-PreferencePanes==2.5.1
+pyobjc-framework-PubSub==2.5.1
+pyobjc-framework-QTKit==2.5.1
+pyobjc-framework-Quartz==2.5.1
+pyobjc-framework-ScreenSaver==2.5.1
+pyobjc-framework-ScriptingBridge==2.5.1
+pyobjc-framework-SearchKit==2.5.1
+pyobjc-framework-ServiceManagement==2.5.1
+pyobjc-framework-Social==2.5.1
+pyobjc-framework-SyncServices==2.5.1
+pyobjc-framework-SystemConfiguration==2.5.1
+pyobjc-framework-WebKit==2.5.1
+pyOpenSSL==19.0.0
+pyparsing==2.0.1
+python-dateutil==1.5
+pytz==2013.7
+requests==2.25.1
+requests-oauthlib==1.3.0
+rsa==4.5
+scipy==0.13.0b1
+six==1.15.0
+uritemplate==3.0.1
+urllib3==1.26.2
+Werkzeug==1.0.1
+xattr==0.6.4
