+++ wsgi.py
@@ -1,0 +1,4 @@
+from sentysis_flask import sentysis
+
+if __name__ == "__main__":
+    sentysis.run()
